open pr

+++ src/components/user/user.controller.ts
@@ -1,4 +1,4 @@
-import { IUser, IUserSelect, IuserUpdate } from './user.interface';
+import { IUser, IUserLogIn, IUserSelect, IuserUpdate } from './user.interface';
 import { Request, Response } from "express";
 import UserService from "./user.service";
 
@@ -17,8 +17,14 @@ import UserService from "./user.service";
         createUser = async (req: Request, res: Response): Promise<any> => {
             const user:IUser = req.body;
             try{
-                const key = await this.userService.createUser(user);
-                res.status(201).send(key);
+                const userExist = await this.userService.getUserByEmail(user.email);
+                if(!userExist.user.length){
+                    const key = await this.userService.createUser(user);
+                    res.status(201).send(key);
+                }else{
+                    res.status(409).send(`[User-magement] Controller : cann't create user, user already exist`);
+                }
+                
             }catch(err){
                 res.status(400).send(`[User-magement] Controller : cann't create user ${err}`);
             }
@@ -47,10 +53,30 @@ import UserService from "./user.service";
         deleteUser = async (req: Request, res: Response): Promise<void> => {
             try{
                 const key:any = req.params;
-                await this.userService.deleteUser(key);
-                res.status(200).send();
+                const userExist = await this.userService.getUserByKey(key);
+                if(!userExist.length){
+                    await this.userService.deleteUser(key);
+                    res.status(200).send();
+                }else{
+                    res.status(409).send(`[User-magement] Controller : cann't delete user, user doesn't exist`);
+                }
             }catch(err){
-
+                res.status(400).send(`[User-magement] Controller : cann't delete user ${err}`);
+            }
+        }
+        
+        login = async (req: Request, res: Response): Promise<void> => {
+            const user:IUserLogIn = req.body;
+            try{
+                const userExist = await this.userService.getUserByEmail(user.email);
+                if(userExist.user.length){
+                    const token = await this.userService.logIn(user);
+                    res.status(201).send(token)
+                }else{
+                    res.status(409).send(`[User-magement] Controller : cann't find user, user doesn't exist`);
+                }
+            }catch(err){
+                res.status(400).send(`[User-magement] Controller : cann't login ${err}`);
             }
         }
     }
